Removing MissedConnectionQuesetion table
There are still remenants of an old data structure that we had. Before,
we had a table called "missed_connection_questions", but we have since
moved that information over to an attribute that lives on
`MissedConnection`. This commit drops that table, and needs to be run in
production:

```
heroku run rake db:migrate
```